LIBOMV-489 Fixes encoding bug in AssetWearable class encoding of Texture UUIDs Patch courtesy of Arthur Valadares
git-svn-id: http://libopenmetaverse.googlecode.com/svn/libopenmetaverse/trunk@2604 52acb1d6-8a22-11de-b505-999d5b087335

+++ OpenMetaverse/AssetTypes.cs
@@ -730,7 +730,7 @@ namespace OpenMetaverse
             data.Append("textures "); data.Append(Textures.Count); data.Append(NL);
             foreach (KeyValuePair<AppearanceManager.TextureIndex, UUID> texture in Textures)
             {
-                data.Append(texture.Key); data.Append(" "); data.Append(texture.Value.ToString()); data.Append(NL);
+                data.Append((byte)texture.Key); data.Append(" "); data.Append(texture.Value.ToString()); data.Append(NL);
             }
 
             AssetData = Utils.StringToBytes(data.ToString());
